backend/requirements.txt
django==5.0.3
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.15.1